HackStudio: Add a widget showing the state of console's running process
We now have a little widget that sits above the terminal view in the
build/application console. When a child process is running, we show its
PID, name, scheduling counter, and amount of resident memory in a live
little overview.

This is not working right just yet, since we don't know how to get to
the actually active PID on the TTY. Or, well, we find the active PID by
looking at the PGID of our fork()ed child.

This manages to find children spawned by Shell, but not children
spawned by make, for instance. I need to figure out how to find those.