Base: Add Tinier Regular, a 5px tall system font
This font is inspired by one I discovered years ago and have been
meaning to add for a long time. It is also a playful jab at @xexxa who
has the previous smallest font :^) I made a few changes to make it fit
in the 5x3 format. The original is licensed CC0 so there is no copyright
infringement.

Original font:
https://robey.lag.net/2010/01/23/tiny-monospace-font.html