Ports: Remove spaces in specifier in 2 printf calls in acpica-tools port
The specifiers are badly written and lead to a crash when using iASL.
This happens because our printf doesn't understand how to handle a space
between the percent symbol and the actual formatted type specifier after
it. To fix this, we just remove the spaces in a new patch file.