LibCompress: Detemplatize generate_huffman_lengths()
Take Spans instead of Arrays. There's no need to have one copy of this
function for every possible array size passed to it.

Hardcode the inline size of the BinaryHeap to 288 for now. If this
becomes a performance issue in the future, we can make that size
an (optional) template parameter then.

No behavior change.