DeviceDeviceMapper: Remove hardcoded list of pluggable once devices
Instead, simplify things by allowing the user to specify a specific
device entry in the configuration file, by specifying a minor number.

The first example of such device is the /dev/beep, as it resides on the
"generic" device node family (as it has a major number of 1).

However, because it can be skipped (if the user disables the beep device
through a kernel commandline option), we can't just create it blindly.
Therefore, when iterating on the configuration file the DeviceMapper
code detects the entry has a specific minor number of 10, and creates
a special DeviceNodeMatch (with a specific minor number being included).
When an event from /dev/devctl that notifies on the existence of the
/dev/beep device arrives, we find this specific match and don't create
an actual DeviceNodeFamily object, but rather blindly create a device
node.Mapper: Remove the concept of pluggable once devices