LibWeb: Add ServiceWorker discarded flag to ESO
Looking at the spec it doesn't seem like there's a chance for a service
worker client to be an environment but not an environment settings
object. In the case that that changes in the implementation, we can
move it.

(cherry picked from commit 237bcbff360d5000ab45da2ca63f196064ab5210)