LibGfx/Path: Implement support for the butt line cap style
This is done by walking half the pen at the two line ends instead
of using each pen vertex (like it happens for round caps).
This won't be completely vertical to the stroke direction, but it
is kind of close, and it's very easy to implement. We'll improve
this later :^)