cat: Allow for multibyte characters when numbering lines
Before, when we tried to use cat -n, if there were unicode characters
we would crash. This was due to us outputting bytes one at a time
using the out function. However, when the out function runs into a
unicode byte on its own, it has no idea what to do with it.