Kernel/aarch64: Support more than 64 interrupts
GICv2s can have up to 1020 interrupts.

AArch64 IRQControllers previously returned a bit field of pending
interrupts. Change the API to return one pending interrupt at a time
instead.

We still truncate the interrupt number to 8 bit in a lot of other places
in the kernel. This is something that needs to be addressed in a future
commit.