LiibGfx/JPEG2000: Reject some images in unsupported color spaces
And add a bunch of test images for this: CMYK, CMYK with alpha,
CIELab, CIELab with alpha, rgba with 16 bits per channel.

I created all of them by changing Image->Mode and then using
Save a Copy… and saving as JPEG2000 lossless, with
"Include Transparency" toggled on or off.

Before this change, most of these images decoded successfully but
looked very wrong, due to interpreting the non-RGB data as RGB.

(The CIELab images have trailing data at the end of the colr box
and are (incorrectly) rejected by
`JPEG2000ColorSpecificationBox::read_from_stream()`.)