Kernel: Clean up the signal trampoline a bit
AArch64 previously calculated the offset to the return value slot
in an incorrect way.

Moving the `pop rcx` after the `mov [rsp+%P1], rax`
makes the offset the same for all architectures,
allowing us to share its calculation.

Also change the order of saving the return value on other
architectures for consistency.