LibGfx/JPEG2000: Combine B.10.7.1 and B.10.7.2 handling
The motivation is that we'll want to make each segment store
its index, and once we add support for the selective arithmetic
coding bypass code block style, we can have single-segment
passes that don't have index 0. We'd then have to be careful to
handle this right in both branches.

Instead, remove the branches. This makes the code look a bit less
like the spec, but that's arguably the spec's fault.

No behavior change.