LibPDF: Implement painting of axial shadings
Or, as they're called elsewhere, linear gradients.

Since this uses PDF function objects and PDF color spaces, it can't use
the existing gradient code in LibGfx.

The way shadings are drawn with the `sh` operator is that one sets a
clip path that's to be filled with the shading and the calls the `sh`
operator.

However, we only support clip rects at the moment, not yet arbitrary
clip paths. So this leads to rectangles with linear gradients
appearing, which sometimes can overlap other page elements. This will
improve once we implement support for arbitrary clip paths.

This is also still missing support for the optional additional /BBox
shading dict entry, which further constrains the clip rect.