LibPDF: Implement basic painting of radial shadings
Or, as they're called elsewhere, radial gradients.

This only implements the basic case without extensions for now,
since that's much simpler.

This requires solving a quadratic equation. We don't have an existing
solver for that yet, so inline a fairly naive one for now. It doesn't
handle catastrophic cancellation case. (It _does_ use the regular
quadratic formula determinant instead of the pq formula determinant,
since that produces noticeably nicer edges in the "two circles with
equal radius" case -- in the lower right of
Tests/LibPDF/shade-radial.pdf.)

This has similar limitations as the axial shading code in that we don't
support arbitrary clip paths yet. For radial shadings, this is slightly
less bad though, since a radial shading that doesn't have its extension
flags set is bound by the outer radius.