Toolchain: Statically link the GNU toolchain
This is done as an effort to support building Serenity on Intel macs,
which runs into link errors when dynamically linking binutil's libctf.

This is a known problem and has been reported on gnu's issue tracker
and on the spack package manager's github issues:
https://lists.nongnu.org/archive/html/bug-binutils/2023-11/msg00050.html
https://github.com/spack/spack/issues/35817
https://github.com/spack/spack/issues/42947

spack also "solves" it by simply *not* dynamically linking.

We even already had an in-tree hack to workaround a different, intl
related, linking error for Intel macs, which this patch removes.