LibGfx+Base: Try harder to add .jpf to known image extensions
I first tried doing this in the second commit of #24016, but
I only updated LibCore/MimeData.cpp there.

I then tried to update Bitmap.h in the first commit of #25671
but gave it type `jpeg` instead of `jpeg2000`. Fix that.
(That commit did successfully update FileTypeFilter.h though!)

And I forgot to update ImageViewer.af, so do this too.

With this, double-clicking a jpf file in File Manager now opens
the image in Image View instead of in Text Editor :^)