LibGfx/JPEG2000: Implement support for progression order 2
With coroutine-based progression iterators, this looks very close
to the spec now :^)

The spec points out that this is an inefficient way to implement
this progression iterator. While that's true, it's still fast
compared to all the other work happening during decoding, so let's
go with the simple approach for now.