Kernel/DeviceTree: Separate drivers into early and non-early drivers
We don't need to probe all drivers this early, so make it possible to
defer probing to init_stage2.

Most subsystems aren't initialized that early, so this also makes it
easier to create and attach devices to their subsystems.

All timer and interrupt controller drivers are now registered with
EARLY_DEVICETREE_DRIVER.

The simple-framebuffer driver is also marked as early so we can use
its framebuffer as the early boot framebuffer.