Kernel/x86: Make the AP setup code work with the EFI prekernel
The long mode AP setup code currently expected 0x2000008??? to be mapped
to 0x8???. This isn't the case if we booted with the EFI prekernel.

We don't need to jump into high memory that early and we can also keep
rbp pointing to the identity mapped AP setup variables, as the
memory at 0x8000 is still identity mapped.