LibPDF: Use read_shading_functions() for shading types 2 and 3
Since types 2 and 3 always use a function, this adds a
`NonemptyShadingFunctionsType` typedef and makes
read_shading_functions() return that.

This has a slight behavior change: read_shading_functions() returns
an error if used with an indexed color space. That's invalid for
types 2 and 3 per

> For similar reasons, an Indexed color space is not allowed in any
> shading whose color values are generated by a function; this rule
> applies to any shading dictionary that contains a Function entry.

but before this change, we didn't check for it. Now we do.