Kernel: Ensure calling process address space restoration in sys$execve
When do_exec() begins to alter address space state, the calling process
could be left in an unstable or incorrect memory context, in the case
of a failure or sys$posix_spawn.

This commit introduces a ScopeGuard at the top of do_exec().
The guard ensures that
`Memory::MemoryManager::enter_process_address_space(Process::current())`
is reliably called upon exiting the scope.