LibGfx: Correctly handle multiple segments in Path::stroke_to_fill()
The last path in svg-stroke-styles.html has three segments. Each is
a triangle, and the first and third segment are closed. The middle,
open triangle is missing its third edge.

For that path, subpath_end_indices contains { 2, 7 } (because the
first triangle has 3 lines since it's closed, the second has 2, and
the third has 3, so the index after the first triangle is 3 - 1 == 2
and the index after the third is (3 + 2 + 3) - 1 == 7).

The existing code was written assuming that both closed and open
segments would have entries in subpath_end_indices. Instead, use
a cursor into subpath_end_indices and increment it only when we hit
a closed segment.

This approach requires filtering out multiple consecutive close
commands. I put that in `Path::segmentize_path()` instead of in
`Path::close()` so that `Path::to_byte_string()` still contains a `'Z'`
for the redundant close commands, but either approach seems valid.

The new `VERIFY()` at the end also requires forcing a new segment for
points after a `ClosePath`, even if the first point on the new
segment is equal to the last point on the old one.

Fixes rendering of the new path in svg-stroke-styles.html and paths.pdf.

gaah bug 1: multiple close