LibPDF: Add Renderer accessors painter() and anti_aliasing_painter()
In the future, we will use different painters sometimes, e.g. when
painting with a clip, or when painting to a transparency group.
This will allow us to swap them out.

(activate_clip() / deactivate_clip() still refer to m_painter because
these methods are not long for this world anyways.)

No behavior change for now.