Tests/LibPDF: Update jbig2-globals.pdf to not contain EndOfPage segment
As detailed in the second commit of #23639, 290c9541deb4, I created
that file by first running

    Meta/jbig2_to_pdf.py -o foo.pdf \
        Tests/LibGfx/test-inputs/jbig2/bitmap.jbig2

But when I created that file, jbig2_to_pdf.py did not yet strip
EndOfPage. It only learned to do that recently, in #26051.

So I ran that command again, then I copied over the new output for the
main jbig2 image, and manually updated the offsets in the xref
table for the two objects following that. I also manually updated
the offset to the xref table in the startxref part.

No real behavior change, but makes jbig2-globals.pdf more
PDF-spec-compliant.