Kernel/aarch64: Build kernel as position-independent code
The EFI Prekernel needs the kernel to be relocatable, so build it
without --no-pie and stop linking the kernel to KERNEL_MAPPING_BASE.

Preprocess the linker script, so we can use PF_{X,W,R} macros instead of
using magic numbers, similar to the linker scripts of other
architectures.

pre_init now needs to relocate itself so it still works when linking
without `--no-pie`.

KSyms.cpp also needs to be changed, as it currently expects the kernel
to be linked to KERNEL_MAPPING_BASE.