LibGfx/JBIG2: Add standard tables F, G, H, K, L, N
This requires adding support for the "lower range table line"
concept, for ranges that start at -infinity up to a given number.
I'm doing this for now by setting the top bit in `prefix_length`.

This is enough to get past 7.4.3.1.2 for both page 1 of annex-h.jbig2
and of page 1 of ignition.pdf.