LibGfx/JPEG: Don't assume that app segments always start with UTF-8
This would otherwise prevent us to decode JPEG files with a seemingly
useless application segment.

The error is triggered by files having an old Photoshop segment with the
signature "Adobe_Photoshop2.5:". As it doesn't end with a null character
we interpret the full segment as the app id, which is obviously not
always UTF-8.

The test case was generated by copying grayscale_app14.jpg and replacing
the null character after APP14's "Adobe" with 0xFF.