LibGfx/JBIG2: Implement TPGRON, and add spec comments for 6.3.5
With this, we can decode 042_24.jb2.

The only way to enter generic_refinement_region_decoding_procedure()
with TPGRON is with a standalone refinement segment. I haven't
encountered them outside of the Power JBIG2 test file set yet. But
they're part of the very first version of JBIG2, so it seems good
to support them.