LibHTTP: Fix a -Wunreachable-code-loop-increment warning
I ran into this while playing with clang's -Wunreachable-code-aggressive
locally.

Before the last commit in #24513, there was a continue somewhere in the
loop and the unconditional break-at-end-of-loop made sense.

After that change, looping until everything was flushed blocked all
downloads.

Rewrite this as an if statement to make the intent more clear.

No behavior change.