Tests/LibGfx: Add JBIG2 test for compositing with 'or', 'xor', 'replace'
The test uses a white background.

To test for 'or', paint a black region (the upper eye) twice in the same
location.

To test for 'xor', paint a black region (the lower eye) once in the same
location, and paint it twice on top of each other at some location that
should not be black.

To test for 'replace', draw a region (everything from pixel x=141 and to
the right) first in inverted, and then again on top with 'replace' in
uninverted.

Preview.app displays this fine.

pdf.js refuses to display a JBIG2 using 'replace' at all.

PDFium treats 'replace' as 'or'.