Tests/LibGfx: Add JBIG2 test for compositing with 'and' and 'xnor'
The test uses a black background.

To test for 'and', paint a region (everything left of x = 141) twice in
the same location.

To test for 'xnor', paint a region (everything at x = 141 and to the
right) once, and paint a smaller region (the lower eye) twice on top of
each other at some location that is meant to be all-white.

Preview.app and PDFium both display this fine.
pdf.js doesn't.