LibGfx/JBIG2+jbig2-from-json: Write intermediate halftone, text regions
Similar to #26410, the challenge with refining halftone and text regions
is that the writer needs to know the decoded halftone and text region
bitmap. That data is easily available in the loader, but not in the
writer.

Similar to #26410, the approach is to call into the *loader* with a
list of segments needed to decode the intermediate region's data.

...and then some minor plumbing to hook up the intermediate region
types in jbig2-from-json.

With this, we can write all region segment types :^)