LibGfx/JBIG2Loader: Set huffman tables in symbol dict decoding procedure
...when calling the text region decoding procedure from it.

These are used when SDREFAGG is set and a REFAGGNINST value for a bitmap
is > 1, and the symbol dictionary segment uses huffman codes.

No behavior change yet because read_symbol_bitmap errors out at the top
when huffman coding is used under these circumstances. But it's easy to
do this part of that, so might as well do it :^)

Also add a note that we are missing tests for these cases.