LibGfx/JBIG2: Add ArithmeticInteger(ID)Encoder
They do what the decoders do, but in reverse.

Add a tests that encode a bunch of values with these and then decode
them with the decoders, and then check we get the same values out again.

This will be used for writing symbol dictionary and text region
segments.

Currently only used in tests, so no behavior change.