LibGfx/JPEGXL: Make the decoding steps independent of the TOC structure
That's something that bugged me for a while as I always had to modify
the two versions when I'm working on this decoder. After loosing time
chasing a bug that was due to us silently skipping the HFGlobal section
in the "single TOC entry" mode, I decided that it was time for me to put
it to a stop.

So this commit is a NFC for files with multiple TOC entries, and we now
actually fail on HFGlobal for single entry files.