Tests/LibGfx: Add test for refine-one in huffman-coded symbol dict
This one only decodes correctly in Acrobat Reader. Similar to the other
huffman refinement tests, it decodes mostly ok in Preview.app with some
artifacts, but the refinement doesn't look right at all. mutool can
decode this better than the two other refinement tests which it can't
decode at all: It also shows the lower eye with artifacts like
Preview.app (but different artifacts) but gets the rest of the file
right.

PDFium cannot decode it, but that will change soon
(crbug.com/459326233). That bug has some more notes on why I think the
file might be correct.

(pdf.js can't decode this either.)

The json file is identical to bitmap-symbol-symhuffrefineone.json,
except that `uses_huffman_encoding` is set to true for both symbol and
text regions (not really needed for the first).