Kernel: Prevent lost wakeups in the idle loop
In the old idle loop implementation, lost wakeups can happen when the
processor receives an interrupt immediately before wait_for_interrupt().
In that situation, we still go to sleep, even if the interrupt caused
a thread to be runnable or we got notified of a runnable thread.

This new idle loop implementation instead runs with interrupts disabled.
The only place in this new idle loop where we can get interrupts is
during Processor::idle(). After that, we will only go to sleep if there
are still no runnable threads.